refactor: divide seed into multi steps
