+++ apps/cms/src/app/next/seed/globals/data.ts
@@ -1,4 +1,4 @@
-import { Header } from "@synoem/types";
+import { Footer, Header } from "@synoem/types";
 
 export const getHeaderData = (
   imageOne: string,
@@ -297,4 +297,130 @@ export const getHeaderData = (
       },
     ],
   };
+};
+
+export const footerData: Omit<Footer, "id" | "updatedAt" | "createdAt"> = {
+  columns: [
+    {
+      blockType: "linkGroup",
+      linkGroup: {
+        title: "Products",
+        links: [
+          {
+            link: {
+              type: "relative",
+              label: "Solar Panels",
+              relative: "/products/solar-panel",
+            },
+          },
+          {
+            link: {
+              type: "relative",
+              label: "Inverters",
+              relative: "/products/inverter",
+            },
+          },
+          {
+            link: {
+              type: "relative",
+              label: "Battery Storage",
+              relative: "/products/battery-storage",
+            },
+          },
+          {
+            link: {
+              type: "relative",
+              label: "Accessories",
+              relative: "/products/accessories",
+            },
+          },
+        ],
+      },
+    },
+
+    {
+      blockType: "linkGroup",
+      linkGroup: {
+        title: "Resources",
+        links: [
+          {
+            link: {
+              type: "relative",
+              label: "Posts",
+              relative: "/posts",
+            },
+          },
+          {
+            link: {
+              type: "relative",
+              label: "Case Studies",
+              relative: "/case-studies",
+            },
+          },
+          {
+            link: {
+              type: "relative",
+              label: "Downlaods",
+              relative: "/downloads",
+            },
+          },
+          {
+            link: {
+              type: "relative",
+              label: "FAQs",
+              relative: "/faqs",
+            },
+          },
+        ],
+      },
+    },
+    {
+      blockType: "linkGroup",
+      linkGroup: {
+        title: "Company",
+        links: [
+          {
+            link: {
+              type: "relative",
+              label: "About Us",
+              relative: "/company/about-us",
+            },
+          },
+          {
+            link: {
+              type: "relative",
+              label: "News & Events",
+              relative: "/company/news",
+            },
+          },
+          {
+            link: {
+              type: "relative",
+              label: "Careers",
+              relative: "/company/careers",
+            },
+          },
+          {
+            link: {
+              type: "relative",
+              label: "Terms of Service",
+              relative: "/terms",
+            },
+          },
+          {
+            link: {
+              type: "relative",
+              label: "Privacy Policy",
+              relative: "/privacy",
+            },
+          },
+        ],
+      },
+    },
+    {
+      blockType: "contactInfo",
+    },
+  ],
+  links: [],
+  copyright: "All rights reserved",
 };
